Streaming synchronization
Convert the output logic into a readable stream. By doing this, the stream can be piped (to stdout).

Add additional messages in the TAP result and fixes some lines that are not standard.

Make sure that in node 0.10 the output buffer is always flushed before exiting. This should prevent some messages from getting lost.